maintenance-system/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
python-jose[cryptography]==3.5.0
bcrypt==4.3.0
python-multipart==0.0.20
pydantic-settings==2.10.1
email-validator==2.3.0
